refactor(welcome): rename app-footer to welcome-footer
Frees the app-footer name for the new mobile-dock and names the
component for what it is — the marketing footer on the welcome page.
