+++ src/components/welcome-footer.vue
